+++ TablePro/Models/Settings/AppSettings.swift
@@ -482,5 +482,15 @@ struct HistorySettings: Codable, Equatable {
 
 /// Tab behavior settings
 struct TabSettings: Codable, Equatable {
+    var enablePreviewTabs: Bool = true
     static let `default` = TabSettings()
+
+    init(enablePreviewTabs: Bool = true) {
+        self.enablePreviewTabs = enablePreviewTabs
+    }
+
+    init(from decoder: Decoder) throws {
+        let container = try decoder.container(keyedBy: CodingKeys.self)
+        enablePreviewTabs = try container.decodeIfPresent(Bool.self, forKey: .enablePreviewTabs) ?? true
+    }
 }